DWT-624 Move shared tests to tests directory
