+++ src/schemas/waste-hazardous.test.js
@@ -3,8 +3,8 @@ import {
   sourceOfComponentsNotProvided,
   validSourceOfComponents
 } from '../common/constants/source-of-components.js'
+import { popsAndHazardousComponentsErrorTests } from '../test/common/pop-and-hazardous-components/pops-and-hazardous-components-error-tests.js'
 import { receiveMovementRequestSchema } from './receipt.js'
-import { testPopsAndHazardousComponentsErrors } from './test-helpers/pops-and-hazardous-components-errors-test-helpers.js'
 import { createTestPayload } from './test-helpers/waste-test-helpers.js'
 
 describe('Receipt Schema Validation - Hazardous', () => {
@@ -196,5 +196,5 @@ describe('Receipt Schema Validation - Hazardous', () => {
     )
   })
 
-  testPopsAndHazardousComponentsErrors('Hazardous')
+  popsAndHazardousComponentsErrorTests('Hazardous')
 })
